requirements.txt
playwright>=1.40.0
boto3>=1.34.0
requests>=2.31.0
strand-agents>=0.1.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0